fix: fix some stuff
