+++ src/components/SearchClue/SearchClue.types.ts
@@ -1,8 +1,7 @@
-import { TInputComponent } from '~components/InputSearch/InputSearch';
+import { TInputComponent } from '~components/InputSearch/InputSearch.types';
 
 export interface ISearchClueProps {
   clearInput?: TInputComponent['clearInput'];
   value: string;
   label: string;
-  setValue: React.Dispatch<React.SetStateAction<string>>;
 }
